refactor(chat): extract assistant turn renderer
