+++ src/routes/Chat/ChatTimeline.tsx
@@ -6,34 +6,30 @@ import type {
   ChatQuestionRequest,
   ChatAttachment,
   ChatMessage,
-  ChatMessagePart,
   TurnOutputRecord,
 } from "../../../electron/chat/common.ts"
 import type { ConnectionProvider } from "../../../electron/connections/common.ts"
 import type { ResolvedArtifactGroup } from "./artifact-resolution.ts"
 import type { AssistantTimelineBlock } from "./assistant-timeline.ts"
-import type { ChatTurn, ChatTurnProcessStatus, ChatTurnRetrySource } from "./chat-turns.ts"
-import type { ProcessOpenPreference } from "./process-activity-open.ts"
+import type { ChatTurn, ChatTurnRetrySource } from "./chat-turns.ts"
 import type { QuestionDraftStore } from "./question-fields.ts"
-import type { TranslateFn } from "@/i18n/i18n"
 import type { ArtifactSelection } from "@/routes/Chat/GeneratedArtifacts"
 import type { TurnOutputSelection } from "@/routes/Chat/TurnOutputs"
 import type { ChatStatus } from "ai"
 import type { StickToBottomContext } from "use-stick-to-bottom"
 
-import { ChevronDown, ChevronRight, ChevronUp } from "lucide-react"
+import { ChevronDown, ChevronUp } from "lucide-react"
 import * as React from "react"
-import { isConnectionlessNoAuthProvider } from "../../../electron/connections/summary.ts"
 import { useArtifactBundles } from "./artifact-bundle-records.ts"
 import { shouldRenderGeneratedArtifactsShelf } from "./artifact-shelf-visibility.ts"
 import { splitAssistantTimelineBlocks, textFromTimelineBlocks } from "./assistant-timeline.ts"
+import { assistantBlockClassName, shouldRenderConnectionSuggestion } from "./assistant-turn-renderer-model.ts"
+import { AssistantBlock, TurnProcessActivity } from "./AssistantTurnRenderer.tsx"
 import { attachmentWithPreview } from "./chat-attachment-utils.ts"
 import {
   activityForChatTurn,
   assistantMessageIdsKey,
-  chatTurnProcessStatus,
   groupChatTurns,
-  isLiveTurnProcess,
   latestAssistantMessage,
   retrySourceFromTurn,
   reuseStableChatTurns,
@@ -43,7 +39,6 @@ import {
   summarizeTurnProcess,
 } from "./chat-turns.ts"
 import { AttachmentList } from "./ChatAttachments.tsx"
-import { ChatErrorNotice } from "./ChatErrorNotice.tsx"
 import {
   AssistantMessageActions,
   ConnectionAuthorizationIssueAction,
@@ -60,13 +55,10 @@ import {
   visibleUserText,
 } from "./message-text.ts"
 import { PermissionRequiredCard } from "./PermissionRequiredCard.tsx"
-import { processOpenAfterStatusChange, processShouldOpenAutomatically } from "./process-activity-open.ts"
 import { QuestionPromptCard } from "./QuestionPromptCard.tsx"
 import { renderBlocks } from "./render-blocks.ts"
-import { formatWholeSecondDuration } from "./tool-activity.ts"
-import { normalizeServiceSlug, toolActionSummary, toolServiceSlug } from "./tool-display.ts"
-import { hasStoppedTool, isActiveToolPart } from "./tool-state.ts"
-import { ToolActivityStep } from "./ToolActivityStep.tsx"
+import { normalizeServiceSlug } from "./tool-display.ts"
+import { hasStoppedTool } from "./tool-state.ts"
 import {
   turnOutputInitialRole,
   turnOutputRecordsByMessageId,
@@ -75,9 +67,7 @@ import {
 } from "./turn-output-records.ts"
 import { TurnOutputShelf } from "./TurnOutputShelf.tsx"
 import { Conversation, ConversationContent, ConversationScrollButton } from "@/components/ai-elements/conversation"
-import { Message, MessageActions, MessageContent, MessageResponse } from "@/components/ai-elements/message"
-import { MarkdownImage } from "@/components/ai-elements/message-image"
-import { Task, TaskContent, TaskTrigger } from "@/components/ai-elements/task"
+import { Message, MessageActions, MessageContent } from "@/components/ai-elements/message"
 import { useT } from "@/i18n/i18n"
 import { cn } from "@/lib/utils"
 
@@ -87,19 +77,6 @@ const GeneratedArtifacts = React.lazy(() =>
 const CHAT_CONTENT_MAX_WIDTH_CLASS = "min-w-0 max-w-[50rem]"
 const ASSISTANT_TEXT_SMOOTH_WINDOW_MS = 45_000
 
-function shouldRenderConnectionSuggestion(
-  authorization: AuthorizationInfo | undefined,
-  providerByService: Map<string, ConnectionProvider>,
-): AuthorizationInfo | undefined {
-  if (!authorization) {
-    return undefined
-  }
-  const provider = providerByService.get(normalizeServiceSlug(authorization.service))
-  if (!provider) {
-    return authorization
-  }
-  return provider.status === "connected" || isConnectionlessNoAuthProvider(provider) ? undefined : authorization
-}
 const EMPTY_ARTIFACT_GROUPS: ResolvedArtifactGroup[] = []
 
 function noopArtifactsAvailable(_selection: ArtifactSelection): void {
@@ -128,364 +105,6 @@ function reuseStableArtifactGroupMap(
     }
   }
   return changed ? stable : previous
-}
-
-function formatSettledToolActivityDuration(parts: ChatMessagePart[]): string | null {
-  let start: number | undefined
-  let end: number | undefined
-  for (const part of parts) {
-    const partStart = part.timing?.start
-    const partEnd = part.timing?.end
-    if (typeof partStart !== "number" || typeof partEnd !== "number" || partEnd < partStart) {
-      continue
-    }
-    start = start === undefined ? partStart : Math.min(start, partStart)
-    end = end === undefined ? partEnd : Math.max(end, partEnd)
-  }
-  return start === undefined || end === undefined ? null : formatWholeSecondDuration(end - start)
-}
-
-function formatProcessDuration(
-  process: ReturnType<typeof summarizeTurnProcess>,
-  now: number,
-  live = false,
-): string | null {
-  const isLive = isLiveTurnProcess(process, live)
-  const toolDuration = !isLive && process.tools.length > 0 ? formatSettledToolActivityDuration(process.tools) : null
-  if (!isLive && toolDuration) {
-    return toolDuration
-  }
-  const start = process.startedAt
-  const end = isLive ? now : process.endedAt
-  if (typeof start !== "number" || typeof end !== "number" || end < start) {
-    return null
-  }
-  return formatWholeSecondDuration(end - start)
-}
-
-function processStatusText(t: TranslateFn, status: ChatTurnProcessStatus): string {
-  switch (status) {
-    case "running":
-      return t("chat.processRunning")
-    case "retrying":
-      return t("chat.processRetrying")
-    case "needsAction":
-      return t("chat.processNeedsAction")
-    case "error":
-      return t("chat.processError")
-    case "stopped":
-      return t("chat.processStopped")
-    case "completed":
-      return t("chat.processCompleted")
-    case "completedWithIssues":
-      return t("chat.processCompletedWithIssues")
-  }
-}
-
-function processTitle(t: TranslateFn, status: ChatTurnProcessStatus, duration: string | null): string {
-  const title = processStatusText(t, status)
-  return duration ? `${title} ${duration}` : title
-}
-
-function TurnProcessActivity({
-  blocks,
-  process,
-  live = false,
-  billingCacheScope,
-  providerByService,
-  onAuthorize,
-  onViewBilling,
-}: {
-  blocks: AssistantTimelineBlock[]
-  process: ReturnType<typeof summarizeTurnProcess>
-  live?: boolean
-  billingCacheScope: string
-  providerByService: Map<string, ConnectionProvider>
-  onAuthorize: (auth: AuthorizationInfo, source?: ChatTurnRetrySource) => void
-  onViewBilling?: () => void
-}) {
-  const t = useT()
-  const status = chatTurnProcessStatus(process, live)
-  const shouldOpen = processShouldOpenAutomatically(status, process.hasFinalAnswer)
-  const statusKey = [
-    status,
-    live ? "live" : "",
-    process.activity?.phase,
-    process.tools.map((part) => `${part.partId}:${part.status}`).join("|"),
-    process.errors.map((part) => part.partId).join("|"),
-  ].join(":")
-  const [open, setOpen] = React.useState(shouldOpen)
-  const [now, setNow] = React.useState(() => Date.now())
-  const duration = formatProcessDuration(process, now, live)
-  const title = processTitle(t, status, duration)
-  const renderBlocks = blocks.map((item) => item.block)
-  const showLiveStatus = renderBlocks.length === 0 && shouldShowLiveStatus(process, status)
-  const titleText = processStatusText(t, status)
-  const activeTool = latestActiveTool(process)
-  const settlingToolPartId =
-    !activeTool && status === "running" && process.activity && process.tools.length > 0
-      ? process.tools.at(-1)?.partId
-      : undefined
-  const openPreferenceRef = React.useRef<ProcessOpenPreference>("auto")
-
-  React.useEffect(() => {
-    setOpen(
-      processOpenAfterStatusChange({
-        hasFinalAnswer: process.hasFinalAnswer,
-        preference: openPreferenceRef.current,
-        status,
-      }),
-    )
-  }, [process.hasFinalAnswer, status, statusKey])
-
-  React.useEffect(() => {
-    if (status !== "running" && status !== "retrying") {
-      return
-    }
-    setNow(Date.now())
-    const timer = window.setInterval(() => setNow(Date.now()), 1000)
-    return () => window.clearInterval(timer)
-  }, [status])
-
-  const handleOpenChange = React.useCallback((nextOpen: boolean) => {
-    openPreferenceRef.current = nextOpen ? "user_open" : "user_closed"
-    setOpen(nextOpen)
-  }, [])
-
-  return (
-    <Task open={open} onOpenChange={handleOpenChange} className="not-prose my-0 w-full">
-      <div className="border-b border-border/60 py-1.5 pr-1.5">
-        <TaskTrigger title={title}>
-          <button
-            type="button"
-            className="group inline-flex max-w-full items-center gap-1.5 text-left font-medium text-[var(--oo-section-heading-foreground)] transition-colors select-none"
-          >
-            <span className="flex min-w-0 items-center gap-1">
-              <span className="min-w-0 truncate">{titleText}</span>
-              {duration ? <span className="shrink-0 tabular-nums">{duration}</span> : null}
-            </span>
-            <ChevronRight className="size-3.5 shrink-0 transition-transform group-data-[state=open]:rotate-90" />
-          </button>
-        </TaskTrigger>
-      </div>
-      <TaskContent className="[&>div]:mt-0">
-        <div className="space-y-2 pt-2">
-          {blocks.map(({ message, block }, index) => (
-            <AssistantBlock
-              key={`${message.id}:${block.kind === "tools" ? block.key : block.part.partId}`}
-              block={block}
-              blockClassName={assistantBlockClassName(renderBlocks, index)}
-              billingCacheScope={billingCacheScope}
-              smoothText={false}
-              providerByService={providerByService}
-              settlingToolPartId={settlingToolPartId}
-              liveTools={live}
-              showAuthorizationPrompt={false}
-              onAuthorize={onAuthorize}
-              onViewBilling={onViewBilling}
-            />
-          ))}
-          {showLiveStatus ? <LiveStatusBar process={process} live={live} /> : null}
-        </div>
-      </TaskContent>
-    </Task>
-  )
-}
-
-function latestActiveTool(process: ReturnType<typeof summarizeTurnProcess>): ChatMessagePart | null {
-  for (let index = process.tools.length - 1; index >= 0; index -= 1) {
-    const part = process.tools[index]
-    if (part && isActiveToolPart(part)) {
-      return part
-    }
-  }
-  return null
-}
-
-function shouldShowLiveStatus(
-  process: ReturnType<typeof summarizeTurnProcess>,
-  status = chatTurnProcessStatus(process),
-): boolean {
-  const activeTool = latestActiveTool(process)
-  return (
-    (status === "running" && !activeTool) ||
-    status === "retrying" ||
-    Boolean(process.activity && status !== "completed" && status !== "stopped")
-  )
-}
-
-function LiveStatusBar({
-  process,
-  live = false,
-}: {
-  process: ReturnType<typeof summarizeTurnProcess> | null
-  live?: boolean
-}) {
-  const t = useT()
-
-  if (!process) {
-    return null
-  }
-
-  const status = chatTurnProcessStatus(process, live)
-  const activeTool = latestActiveTool(process)
-  if (!shouldShowLiveStatus(process, status)) {
-    return null
-  }
-
-  const text = (() => {
-    if (status === "retrying" && process.activity) {
-      return activityText(t, process.activity)
-    }
-    if (activeTool) {
-      return t("chat.liveStatusTool", { action: toolActionSummary(t, activeTool) })
-    }
-    if (process.activity) {
-      return activityText(t, process.activity)
-    }
-    return processTitle(t, status, null)
-  })()
-
-  return (
-    <div className="rounded-md text-muted-foreground">
-      <div className="flex min-h-6 min-w-0 items-center">
-        <LoadingShimmerText className="min-w-0 truncate">{text}</LoadingShimmerText>
-      </div>
-    </div>
-  )
-}
-
-function activityText(t: TranslateFn, activity: AssistantActivityEvent | null): string {
-  switch (activity?.phase) {
-    case "retrying":
-      return activity.attempt
-        ? t("chat.activityRetryingWithAttempt", { attempt: activity.attempt })
-        : t("chat.activityRetrying")
-    case "finalizing":
-      return t("chat.activityFinalizing")
-    case "thinking":
-    default:
-      return t("chat.activityThinking")
-  }
-}
-
-function statusPartText(t: TranslateFn, part: ChatMessagePart): string {
-  switch (part.statusType) {
-    case "reconnecting":
-      return part.attempt && part.maxAttempts
-        ? t("chat.connectionReconnectingWithAttempt", { attempt: part.attempt, maxAttempts: part.maxAttempts })
-        : t("chat.connectionReconnecting")
-    case "reconnected":
-      return t("chat.connectionReconnected")
-    case "connectionFailed":
-      return t("chat.connectionFailed")
-    case "generationStale":
-      return t("chat.generationStale")
-    case "toolRunningWithoutOutput":
-      return t("chat.toolRunningWithoutOutput")
-    case "runtimeRestarting":
-      return part.attempt && part.maxAttempts
-        ? t("chat.runtimeRestartingWithAttempt", { attempt: part.attempt, maxAttempts: part.maxAttempts })
-        : t("chat.runtimeRestarting")
-    case "runtimeRecovered":
-      return t("chat.runtimeRecovered")
-    case "runtimeFailed":
-      return t("chat.runtimeFailed")
-    default:
-      return part.text ?? ""
-  }
-}
-
-type AssistantBlockType = ReturnType<typeof renderBlocks>[number]
-
-function assistantBlockClassName(blocks: AssistantBlockType[], index: number): string | undefined {
-  if (index === 0) {
-    return undefined
-  }
-  const previous = blocks[index - 1]
-  const current = blocks[index]
-  if (!previous || !current) {
-    return undefined
-  }
-  if (previous.kind === "tools" && current.kind === "tools") {
-    return "mt-1"
-  }
-  if (previous.kind !== current.kind) {
-    return "mt-3"
-  }
-  return "mt-2"
-}
-
-function AssistantBlock({
-  block,
-  blockClassName,
-  billingCacheScope,
-  smoothText,
-  providerByService,
-  settlingToolPartId,
-  liveTools = true,
-  showAuthorizationPrompt = true,
-  onAuthorize,
-  onViewBilling,
-}: {
-  block: AssistantBlockType
-  blockClassName?: string
-  billingCacheScope: string
-  smoothText: boolean
-  providerByService: Map<string, ConnectionProvider>
-  settlingToolPartId?: string
-  liveTools?: boolean
-  showAuthorizationPrompt?: boolean
-  onAuthorize: (auth: AuthorizationInfo, source?: ChatTurnRetrySource) => void
-  onViewBilling?: () => void
-}) {
-  const t = useT()
-  return (
-    <div className={blockClassName}>
-      {block.kind === "text" ? (
-        block.part.text ? (
-          <MessageResponse smooth={smoothText}>{block.part.text}</MessageResponse>
-        ) : null
-      ) : block.kind === "error" ? (
-        <ChatErrorNotice
-          autoOpenKey={block.part.partId}
-          billingCacheScope={billingCacheScope}
-          errorCode={block.part.errorCode}
-          errorKind={block.part.errorKind}
-          message={block.part.errorText ?? block.part.error ?? t("chatError.failed.description")}
-          onViewBilling={onViewBilling}
-        />
-      ) : block.kind === "status" ? (
-        <div className="text-sm leading-6 font-medium text-muted-foreground/80">{statusPartText(t, block.part)}</div>
-      ) : block.kind === "attachment" ? (
-        block.part.attachment ? (
-          <AssistantAttachment attachment={block.part.attachment} />
-        ) : null
-      ) : (
-        <div className="space-y-0.5">
-          {block.parts.map((part) => {
-            const service = toolServiceSlug(part)
-            return (
-              <ToolActivityStep
-                key={part.partId}
-                part={part}
-                provider={service ? providerByService.get(service) : undefined}
-                live={liveTools}
-                shimmer={part.partId === settlingToolPartId}
-                settling={part.partId === settlingToolPartId}
-                showAuthorizationPrompt={showAuthorizationPrompt}
-                onAuthorize={onAuthorize}
-              />
-            )
-          })}
-        </div>
-      )}
-    </div>
-  )
-}
-
-function AssistantAttachment({ attachment }: { attachment: ChatAttachment }) {
-  return <MarkdownImage src={attachment.path} alt={attachment.name} />
 }
 
 function MessageBubble({
